refactor: MusclePart interface
