+++ src/main/java/stu_109601003/finalproject/Parts/ARM.java
@@ -1,13 +1,13 @@
-package stu_109601003.finalproject.Handlers;
+package stu_109601003.finalproject.Parts;
 
-public enum ARMPART {
+public enum ARM implements MusclePart {
     ARM("手臂"),
     BICEPS("肱二頭"),
     TRICEPS("肱三頭");
 
     private final String name;
 
-    ARMPART(String name) {
+    ARM(String name) {
         this.name = name;
     }
 
@@ -15,3 +15,4 @@ public enum ARMPART {
         return name;
     }
 }
+
