+++ src/main/java/stu_109601003/finalproject/Parts/MusclePart.java
@@ -1,0 +1,5 @@
+package stu_109601003.finalproject.Parts;
+
+public interface MusclePart {
+    String getName();
+}
